added board print to uci
