+++ src/state/board_info.h
@@ -45,4 +45,6 @@ Bitboard_t AllD12Sliders(BoardInfo_t* boardInfo, Color_t color);
 
 Color_t ColorOfPiece(BoardInfo_t* boardInfo, Square_t square);
 
+void PrintChessboard(BoardInfo_t* info);
+
 #endif
