+++ tdd/debug.h
@@ -29,8 +29,6 @@ void PrintBitboard(Bitboard_t b);
 
 void PrintMailbox(BoardInfo_t *info);
 
-void PrintChessboard(BoardInfo_t* info);
-
 void PrintMoveList(MoveEntryList_t* moveList, BoardInfo_t* info);
 
 void PrintMove(Move_t move, bool hasNewline);
